# add resync controller 