The related feature request: open-cluster-management-io/community#52
Ref: https://github.com/open-cluster-management/backlog/issues/18210

Signed-off-by: haoqing0110 <qhao@redhat.com>

In the benchemark testing, then resync interval is 5 seconds, the schedule time of 100000 placements reduce about 20s. 

Before (resync in the same controller with [code](https://github.com/haoqing0110/placement/commit/546eefb94d7621a700677f98a815f66fc33771be))
```
# go test -bench="BenchmarkSchedulePlacements100000$"
goos: linux
goarch: amd64
pkg: open-cluster-management.io/placement/test/benchmark
cpu: Intel Core Processor (Haswell, no TSX)
BenchmarkSchedulePlacements100000-24                   1        279440555550 ns/op
PASS
ok      open-cluster-management.io/placement/test/benchmark     279.802s
-
```

After  (resync in the new controller)
```
#  go test -bench="BenchmarkSchedulePlacements100000$"
goos: linux
goarch: amd64
pkg: open-cluster-management.io/placement/test/benchmark
cpu: Intel Core Processor (Haswell, no TSX)
BenchmarkSchedulePlacements100000-24                   1        255090872924 ns/op
PASS
ok      open-cluster-management.io/placement/test/benchmark     255.366s
```