@haoqing0110 2022-01-12
@qiujian16 any suggestion on the benchmark testing? don't see a big improvement when adding another controller.