@qiujian16 2022-01-13
what is the reconcile time for each placement? How many clusters in your test. I think we will have to consider parallel