@haoqing0110 2022-01-18
Update the benchmark testing, create 1000/10000 placement in env with 1000 clusters and 1000 addon scores. Below are the total time to schedule all the placements:

Resync in the same controller (with code [here](https://github.com/haoqing0110/placement/commit/1a54b1ab8ba5569f645198aee1fbbec9fb3f80f5))
```
# go test -bench="BenchmarkSchedulePlacements1000$"
goos: linux
goarch: amd64
pkg: open-cluster-management.io/placement/test/benchmark
cpu: Intel Core Processor (Haswell, no TSX)
BenchmarkSchedulePlacements1000-24             1        14613978548 ns/op
PASS
I0118 08:28:45.375109 3827872 base_controller.go:166] Shutting down SchedulingController ...
ok      open-cluster-management.io/placement/test/benchmark     30.056s
```
```
# go test -bench="BenchmarkSchedulePlacements10000$"
goos: linux
goarch: amd64
pkg: open-cluster-management.io/placement/test/benchmark
cpu: Intel Core Processor (Haswell, no TSX)
BenchmarkSchedulePlacements10000-24            1        306247209814 ns/op
PASS
I0118 08:41:38.465052 3830250 base_controller.go:166] Shutting down SchedulingController ...
ok      open-cluster-management.io/placement/test/benchmark     321.011s
```
Resync in the new controller (with this PR)
```
# go test -bench="BenchmarkSchedulePlacements1000$"
goos: linux
goarch: amd64
pkg: open-cluster-management.io/placement/test/benchmark
cpu: Intel Core Processor (Haswell, no TSX)
BenchmarkSchedulePlacements1000-24             1        12399352374 ns/op
PASS
I0118 08:19:39.180974 3824050 base_controller.go:166] Shutting down SchedulingControllerResync ...
ok      open-cluster-management.io/placement/test/benchmark     26.600s
```
```
# go test -bench="BenchmarkSchedulePlacements10000$"
goos: linux
goarch: amd64
I0118 08:25:41.722598 3824874 base_controller.go:166] Shutting down SchedulingController ...
pkg: open-cluster-management.io/placement/test/benchmark
cpu: Intel Core Processor (Haswell, no TSX)
BenchmarkSchedulePlacements10000-24            1        273870701609 ns/op
PASS
I0118 08:25:41.722711 3824874 base_controller.go:166] Shutting down SchedulingControllerResync ...
ok      open-cluster-management.io/placement/test/benchmark     289.023s
```